+++ src/genai/summarize.py
